+++ src/user/termsOfUse/TermsOfUse.less
@@ -1,4 +1,4 @@
-@import "../variables";
+@import "../../variables";
 
 .TermsOfUse__error {
   margin-bottom: 2rem;
@@ -27,6 +27,11 @@
 
 .TermsOfUse__buttons .Button:not(:last-child) {
   margin-right: @button-marginRight;
+}
+
+.TermsOfUse__spinner {
+  display: flex;
+  justify-content: center;
 }
 
 @media (max-width: @screen-xs-max) {
